EPML-5601: Handled the timeout scenario occurred due to wrong type cast.
git-tfs-id: [https://tfs.epmlive.com/tfs/EPMLive/]$/EPMLive2013Release/Main;C7285

+++ PortfolioEngineCore/PortfolioEngineCore/Admininfos/AdminFunctions.cs
@@ -210,7 +210,7 @@ namespace PortfolioEngineCore
                         {
                             int bcuid = DBAccess.ReadIntValue(row["BA_BC_UID"]);
                             int prdid = DBAccess.ReadIntValue(row["BA_PRD_ID"]);
-                            double oldrate = DBAccess.ReadIntValue(row["BA_RATE"]);
+                            double oldrate = DBAccess.ReadDoubleValue(row["BA_RATE"]);
 
                             if (periodrates.ContainsKey(prdid))
                             {
